NODE-1115: Move few functions to utils.py
