+++ integration-testing/client/CasperLabsClient/casperlabs_client/utils.py
@@ -1,0 +1,26 @@
+import os
+import ssl
+import pkg_resources
+import google.protobuf.text_format
+
+
+def hexify(o):
+    """
+    Convert protobuf message to text format with cryptographic keys and signatures in base 16.
+    """
+    return google.protobuf.text_format.MessageToString(o)
+
+
+def bundled_contract(file_name):
+    """
+    Return path to contract file bundled with the package.
+    """
+    p = pkg_resources.resource_filename(__name__, file_name)
+    if not os.path.exists(p):
+        raise Exception(f"Missing bundled contract {file_name} ({p})")
+    return p
+
+
+def extract_common_name(certificate_file: str) -> str:
+    cert_dict = ssl._ssl._test_decode_cert(certificate_file)
+    return [t[0][1] for t in cert_dict["subject"] if t[0][0] == "commonName"][0]
