+++ integration-testing/test/test_check_deploy_signatures.py
@@ -14,7 +14,7 @@ from casperlabs_local_net.grpc_proxy import (
     block_justification,
     update_hashes_and_signature,
 )
-from casperlabs_client import hexify, extract_common_name
+from casperlabs_client.utils import hexify, extract_common_name
 
 
 class GenerateBlockWithNoSignaturesGossipInterceptor(grpc_proxy.GossipInterceptor):
