+++ integration-testing/test/test_equivocation.py
@@ -13,7 +13,7 @@ from casperlabs_local_net.grpc_proxy import (
     block_justification,
     update_hashes_and_signature,
 )
-from casperlabs_client import hexify, extract_common_name
+from casperlabs_client.utils import hexify, extract_common_name
 
 
 class GenerateEquivocatingBlocksGossipInterceptor(grpc_proxy.GossipInterceptor):
